Move PackageSelect and seperate util functions
